testsuite: merge tests for shape based constraints.
