+++ testsuite/CMakeLists.txt
@@ -49,9 +49,7 @@ endfunction(PYTHON_TEST)
 
 python_test(FILE cellsystem.py MAX_NUM_PROC 4)
 python_test(FILE constraint_homogeneous_magnetic_field.py MAX_NUM_PROC 4)
-python_test(FILE constraint_shape_based_cylinder.py MAX_NUM_PROC 4)
-python_test(FILE constraint_shape_based_ellipsoid.py MAX_NUM_PROC 4)
-python_test(FILE constraint_shape_based_wall.py MAX_NUM_PROC 4)
+python_test(FILE constraint_shape_based.py MAX_NUM_PROC 4)
 python_test(FILE coulomb_cloud_wall.py MAX_NUM_PROC 4)
 python_test(FILE coulomb_tuning.py MAX_NUM_PROC 4)
 python_test(FILE correlation.py MAX_NUM_PROC 4)
